refactor setup package again
